refac: update variables and routes
