+++ frontend-portal/src/index.js
@@ -15,6 +15,7 @@ export default function Index(){
       <Routes>
         < Route path="/" element={<App/>}/>
         < Route path="/show" element={<ShowComponent/>}/>
+        
       </Routes>
     </BrowserRouter>
 
@@ -23,8 +24,4 @@ export default function Index(){
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render( <Index />);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
